Added logical expressions. Completed all three tasks (not sure). Edited Data.java class, Lexer.jflex file and finally Parser.cup now can work with logical_expressions.

+++ src/main/java/ru/omsu/translator/cup/Parser.java
@@ -362,7 +362,7 @@ class CUP$Parser$actions {
         CustomSymbol sym = (CustomSymbol)id;
         Token token = sym.getToken();
         String varName = (String)token.getValue();
-        RESULT = Data.getVariable(varName);
+        RESULT = (Integer) Data.getVariable(varName);
     
               CUP$Parser$result = parser.getSymbolFactory().newSymbol("factor",5, ((java_cup.runtime.Symbol)CUP$Parser$stack.peek()), ((java_cup.runtime.Symbol)CUP$Parser$stack.peek()), RESULT);
             }
